implementing xnet test
